#552: move Category tree building code to IPEDSource class
