+++ iped-engine/src/main/java/gpinf/dev/data/Category.java
@@ -89,10 +89,6 @@ public class Category implements Serializable, Comparable<Category> {
         return -1;
     }
 
-    public void clearItemCount() {
-        numItems = -1;
-    }
-
     @Override
     public int compareTo(Category o) {
         return collator.compare(CategoryLocalization.getInstance().getLocalizedCategory(name),
@@ -108,6 +104,7 @@ public class Category implements Serializable, Comparable<Category> {
         if (this.parent == null)
             return name;
         String name = CategoryLocalization.getInstance().getLocalizedCategory(this.name);
+        name = adjustCase(name);
         if (numItems == -1) {
             return name + " (...)"; //$NON-NLS-1$
         } else {
@@ -126,6 +123,18 @@ public class Category implements Serializable, Comparable<Category> {
             clone.children.add(newChild);
         }
         return clone;
+    }
+
+    private String adjustCase(String cat) {
+        StringBuilder str = new StringBuilder();
+        for (String s : cat.split(" ")) //$NON-NLS-1$
+            if (s.length() == 3)
+                str.append(s.toUpperCase() + " "); //$NON-NLS-1$
+            else if (s.length() > 3)
+                str.append(s.substring(0, 1).toUpperCase() + s.substring(1) + " "); //$NON-NLS-1$
+            else
+                str.append(s + " "); //$NON-NLS-1$
+        return str.toString().trim();
     }
 
 }
